requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
Flask==2.1.2
future==0.18.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
ratelim==0.1.6
requests==2.28.0
six==1.16.0
twilio==7.9.3
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
